create reservation

+++ HotelReservationSystem/Models/ReservationManagement/Reservation.cs
@@ -8,7 +8,7 @@ public class Reservation : BaseModel
     public ReservationStatus ReservationStatus { get; set; }
     public int NumberOfGuests { get; set; }
     public string SpecialRequests { get; set; }
-    public decimal TotalAmount { get; set; }
+    public double TotalAmount { get; set; }
     public int GuestID { get; set; }
     public Guest Guest { get; set; }
     public ICollection<ReservationRoom> ReservationRooms { get; set; }
