+++ HotelReservationSystem/ViewModels/ReservationRooms/AvailableRoomViewModel.cs
@@ -7,7 +7,7 @@ public class AvailableRoomViewModel
     public int RoomID { get; set; }
     public string RoomType { get; set; }
     public string RoomNumber { get; set; }
-    public decimal? BasicPrice { get; set; }
+    public double? BasicPrice { get; set; }
     
     public IEnumerable<FacilityViewModel> Facilities { get; set; }
 }
